Extract browser actions runner
